requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
SQLAlchemy>=2.0
python-dotenv
openai>=1.0.0
numpy
python-dateutil
elasticsearch>=8.13.0,<9.0.0
httpx[socks]>=0.27.0
